+++ src/app/login/page.tsx
@@ -1,0 +1,12 @@
+import SignIn from "@/components/SignIn"
+import { Box } from "@mui/material"
+
+
+export default function LoginPage() {
+
+  return (
+    <Box>
+        <SignIn/>
+    </Box>
+  )
+}
