+++ src/app/signup/page.tsx
@@ -1,0 +1,12 @@
+import SignUp from "@/components/SignUp"
+import { Box } from "@mui/material"
+
+
+export default function SignupPage() {
+
+  return (
+    <Box>
+        <SignUp/>
+    </Box>
+  )
+}
