+++ src/components/icons/SitemarkIcon.tsx
@@ -1,3 +1,5 @@
+import * as React from 'react';
+import SvgIcon from '@mui/material/SvgIcon';
 
 export default function SitemarkIcon() {
   return (
