+++ src/theme/ColorModeIconDropdown.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import * as React from 'react';
 import DarkModeIcon from '@mui/icons-material/DarkModeRounded';
 import LightModeIcon from '@mui/icons-material/LightModeRounded';
